feat: set up handler for assignment , remove all typescript errors
