+++ exam_frontend/src/components/AssignAsset/Assign/Assignment.tsx
@@ -1,115 +1,37 @@
-import axiosInstance from '../../../config/AxiosConfig';
-import { useQuery, useMutation } from '@tanstack/react-query';
 import styles from './Assignment.module.css';
 import { ApiResponse, EmployeeDetails } from './types';
-import { useState, useEffect} from 'react';
-import { DataType } from '../AssetTable/types';
-import { message } from 'antd';
+import { DataType } from '../../AssetTable/types';
+
 
 
 interface AssignmentProps {
-  record: DataType;
+  value: string;
+  employeeId: number | undefined;
+  divVisible: boolean;
+  employeeDepartment: string;
+  employeeDesignation: string;
+  employeeName: string;
+  data: ApiResponse | undefined;
+  handleInputChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  handleNameClick: (name: string, id: number, department: string, designation: string) => void;
+  handleAssign: () => void;
+  record:DataType | null;
 }
 
-export const Assignment: React.FC<AssignmentProps> = ({ record }) => {
-  const [query, setQuery] = useState<string>("");
-  const [value, setValue] = useState<string>("");
-  const [fetchData, setFetchData] = useState<boolean>(false); // Initialize as false
-  const [employeeId, setEmployeeId] = useState<number>();
-  const [divVisible,setdivVisible] = useState<boolean>(false)
-  const [employeeDepartment,setEmployeeDepartment]=useState<string>("")
-  const [employeeDesignation,setEmployeeDesignation] = useState<string>("")
-  const [employeeName,setEmployeeName] = useState<string>("")
-
-  const { data, isLoading, isError } = useQuery<ApiResponse>({
-
-    queryKey: ['Assign'],
-    enabled: fetchData && query.trim().length > 0,
-    queryFn: (): Promise<ApiResponse> => axiosInstance.get(`/asset/employee?name=${query}`).then((res) => {
-      console.log(res);
-      console.log("res.data", res.data.data);
-      return res.data;
-    }),
-    onSuccess: () => {
-     console.log("success")
-     setFetchData(false)
-    },
-    onError: () => {
-      // Reset the states in case of error
-      setValue("");
-      setEmployeeId(undefined);
-    }
-  });
-
-  const mutation = useMutation(
-    (requestData: any) => axiosInstance.post('/asset/assign_asset', requestData),
-    {
-      onSuccess: () => {
-        message.success("successfully assigned")
-      },
-      onError: (error) => {
-        message.error("unsuccessful, the asset may be already assigned")
-        console.log(error)
-      }
-    }
-  );
-
-  useEffect(() => {
-    
-    if(value=="")
-    {
-      setEmployeeId(undefined)
-      
-
-    }
-    return () => {
-      
-    };
-  }, [value]);
-
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    console.log(value,"value",name,"name")
-    setdivVisible(true)
-    const inputValue = event.target.value;
-    console.log("in the handleinputchange " ,inputValue.trim().length)
-    setQuery(inputValue);
-    setValue(inputValue);
-    if (inputValue.trim().length > 0) {
-      console.log("input value > 0",inputValue)
-      setFetchData(true);
-    }
-    if(value!=employeeName)
-     setEmployeeId(undefined)
-  };
-
-  const handleNameClick = (name: string, id: number, department:string, designation:string,) => {
-    setValue(name);
-    setEmployeeName(name)
-    
-    setEmployeeId(id);
-    setdivVisible(false)
-    setEmployeeDepartment(department)
-    setEmployeeDesignation(designation)
-    console.log("handle name clck worked")
-    
-  };
-
-  const handleAssign = () => {
-    setValue("")
+export const Assignment: React.FC<AssignmentProps> = ({
+  value,
+  employeeId,
+  divVisible,
+  employeeDepartment,
+  employeeDesignation,
+  employeeName,
+  data,
+  handleInputChange,
+  handleNameClick,
+  handleAssign,
+  record,
+}) => {
   
- 
-    if (data != null && !record.custodian) {
-      const requestBody = {
-        id: employeeId,
-        asset_uuid: record.key
-      };
-      console.log(requestBody);
-      mutation.mutate(requestBody);
-    } else {
-      alert("Please select an employee");
-    }
-  };
-
 
   
 
@@ -121,43 +43,43 @@ export const Assignment: React.FC<AssignmentProps> = ({ record }) => {
     <div className="grid font-display grid-cols-2 gap-3 lg:grid-cols-5 my-3 text-sm">
       <div>
         <label htmlFor="productName">ASSET NAME:</label>
-        <input type="text" id="productName" value={record.product_name} className='mt-1 font-display rounded-lg'  disabled/>
+        <input type="text" id="productName" value={ record?.product_name} className='mt-1 font-display rounded-lg'  disabled/>
       </div>
       <div>
         <label htmlFor="assetType">ASSET TYPE:</label>
-        <input type="text" id="assetType" value={record.asset_type} className='mt-1 font-display rounded-lg' disabled/>
+        <input type="text" id="assetType" value={record?.asset_type} className='mt-1 font-display rounded-lg' disabled/>
       </div>
       <div>
         <label htmlFor="configuration">CONFIGURATION:</label>
-        <input type="text" id="configuration" value={record.configuration} className='mt-1 font-display rounded-lg'  disabled/>
+        <input type="text" id="configuration" value={record?.configuration} className='mt-1 font-display rounded-lg'  disabled/>
       </div>  
     <div>
       <label htmlFor="location">LOCATION:</label>
-      <input type="text" id="location" value={record.location} className='mt-1 font-display rounded-lg' disabled />
+      <input type="text" id="location" value={record?.location} className='mt-1 font-display rounded-lg' disabled />
     </div>
     <div>
       <label htmlFor="serialNumber">ASSET SERIAL NUMBER:</label>
-      <input type="text" id="serialNumber" value={record.serial_number} className='mt-1 font-display rounded-lg' disabled/>
+      <input type="text" id="serialNumber" value={record?.serial_number} className='mt-1 font-display rounded-lg' disabled/>
     </div>
     <div>
       <label htmlFor="modelNumber">ASSET MODEL NUMBER:</label>
-      <input type="text" id="modelNumber" value={record.model_number} className='mt-1 font-display rounded-lg' disabled />
+      <input type="text" id="modelNumber" value={record?.model_number} className='mt-1 font-display rounded-lg' disabled />
     </div>
     <div>
       <label htmlFor="version">VERSION:</label>
-      <input type="text" id="version" value={record.version} className='mt-1 font-display rounded-lg' disabled />
+      <input type="text" id="version" value={record?.version} className='mt-1 font-display rounded-lg' disabled />
     </div>
     <div>
       <label htmlFor="version">VERSION:</label>
-      <input type="text" id="version" value={record.version} className='mt-1 font-display rounded-lg' disabled />
+      <input type="text" id="version" value={record?.version} className='mt-1 font-display rounded-lg' disabled />
     </div>
     <div>
       <label htmlFor="version">VERSION:</label>
-      <input type="text" id="version" value={record.version} className='mt-1 font-display rounded-lg' disabled />
+      <input type="text" id="version" value={record?.version} className='mt-1 font-display rounded-lg' disabled />
     </div>
     <div>
       <label htmlFor="version">VERSION:</label>
-      <input type="text" id="version" value={record.version} className='mt-1 font-display rounded-lg' disabled />
+      <input type="text" id="version" value={record?.version} className='mt-1 font-display rounded-lg' disabled />
     </div>
     </div>
   </form>
